invoice_no candidates fix and candidate extraction done

+++ utils/generate_tesseract_results.py
@@ -3,13 +3,19 @@ from pathlib import Path
 import cv2
 import json
 import pytesseract
-imgs = list(Path('ALL').glob("*.jpg"))
-for img in imgs:
+import os
+from tqdm import tqdm
+
+
+imgs = list(Path('../dataset/ALL').glob("*.jpg"))
+for img in tqdm(imgs[950:]):
     try:
+        if os.path.exists('../dataset/tesseract_results/'+str(img.stem) + '.json'):
+            continue
         image = cv2.imread(str(img))
-        d = pytesseract.image_to_data(image, output_type=Output.DICT)
-        with open('tesseract_results/'+str(img.stem) +'.json','w') as f:
-            json.dump(d,f)
+        d = pytesseract.image_to_data(image, config='--oem 1 --psm 3', output_type=Output.DICT)
+        with open('../dataset/tesseract_results/'+str(img.stem) + '.json', 'w') as f:
+            json.dump(d, f)
     except Exception as e:
         print(img,e)
         continue
